Remove duplicated code from NetworkManagerClient
